refactor(mobile): extract  import and export functionality for data setting
